C#: Make FeedManager unit-testable.
